Оптимизировал компонент MainScreen, выделил зону со списком карточек и картой в отдельный компонент ContentArea, чтобы действия с карточками не приводили к перерисовке всего MainScreen. Добавил компонент EmptyContentArea, который отрисовывается в случае, если для выбранного города нет доступных объявлений
