+++ project/src/components/main-screen/main-screen.tsx
@@ -1,31 +1,14 @@
-import OffersList from '../offers-list/offers-list';
 import Header from '../header/header';
-import Map from '../map/map';
-import Sort from '../sort/sort';
-import { useState } from 'react';
 import CitiesList from '../cities-list/cities-list';
 import { getCurrentCity} from '../../store/main-screen/selectors';
-import { getSortedOffers } from '../../store/data-offers/selectors';
+import { getOffersInCurrentCity } from '../../store/data-offers/selectors';
 import { useSelector} from 'react-redux';
-
-const getCitiesCoordinates = (city:string) => {
-  switch(city){
-    case 'Paris':{
-      return {latitude: 48.8534,longitude: 2.3488,zoom: 10};
-    }
-    case 'Amsterdam':{
-      return {latitude: 52.3909553943508,longitude: 4.85309666406198,zoom: 10};
-    }
-    default :{
-      return {latitude: 52.3909553943508,longitude: 4.85309666406198,zoom: 10};
-    }
-  }
-};
+import ContentArea from '../content-area/content-area';
+import EmptyContentArea from '../empty-content-area/empty-content-area';
 
 function MainScreen(): JSX.Element {
-  const [idActiveOffer, setIdActiveOffer] = useState<null | number>(null);
   const currentCity = useSelector(getCurrentCity);
-  const sortedOffers = useSelector(getSortedOffers);
+  const OffersInCurrentCity = useSelector(getOffersInCurrentCity);
 
   return (
     <div className="page page--gray page--main">
@@ -36,17 +19,11 @@ function MainScreen(): JSX.Element {
           <CitiesList currentCity={currentCity}/>
         </div>
         <div className="cities">
-          <div className="cities__places-container container">
-            <section className="cities__places places">
-              <h2 className="visually-hidden">Places</h2>
-              <b className="places__found">{sortedOffers.length} places to stay in {currentCity}</b>
-              <Sort />
-              <OffersList offers={sortedOffers} onCardFocus={setIdActiveOffer}/>
-            </section>
-            <div className="cities__right-section">
-              <Map city={getCitiesCoordinates(currentCity)} offers={sortedOffers} idActiveOffer={idActiveOffer}/>
-            </div>
-          </div>
+          {
+            OffersInCurrentCity.length > 0
+              ? <ContentArea currentCity={currentCity} />
+              : <EmptyContentArea currentCity={currentCity} />
+          }
         </div>
       </main>
     </div>);
